Separate branch type to opcode logic
